requirements.txt
Django==2.1.5
django-debug-toolbar==1.11
django-rest-swagger==2.2.0
djangorestframework==3.9.1
psycopg2-binary==2.7.6.1